% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedWald.R
\name{adjustedWald}
\alias{adjustedWald}
\title{Adjusted Wald confidence interval for binomial distributions}
\usage{
adjustedWald(s, n, l)
}
\arguments{
\item{s}{success count}

\item{n}{sample size}

\item{l}{confifence level given as decimal, e.g. 95\% is 0.95}
}
\value{
numeric vector of interval of length 2
}
\description{
Adjusted Wald confidence interval for binomial distributions
}
