% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_continuous.R
\name{benchmark_continuous}
\alias{benchmark_continuous}
\title{Benchmark test for continuous data, e.g. satisfaction scores}
\usage{
benchmark_continuous(m, mu, std, n, ci = FALSE, alpha = 0.1)
}
\arguments{
\item{m}{sample mean}

\item{mu}{target benchmark}

\item{std}{sample standard deviation}

\item{n}{sample size}

\item{ci}{Boolean, if TRUE, confidence interval is also reported}

\item{alpha}{for confidence interval}
}
\description{
Benchmark test for continuous data, e.g. satisfaction scores
}
