% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_log.R
\name{benchmark_log}
\alias{benchmark_log}
\title{Benchmark test for task times (positively skewed, long right tail data)}
\usage{
benchmark_log(x, mu, ci = FALSE, alpha = 0.1)
}
\arguments{
\item{x}{data, numeric vector}

\item{mu}{target benchmark}

\item{ci}{Boolean, if TRUE, confidence interval is also reported alongside the geometric mean (vector length of 4)}

\item{alpha}{for confidence interval}
}
\description{
Benchmark test for task times (positively skewed, long right tail data)
}
