% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_test_large_sample.R
\name{benchmark_test_large_sample}
\alias{benchmark_test_large_sample}
\title{Benchmark test for large sample
Small sample: success and failure counts are at least 15
Performs a one-sided test}
\usage{
benchmark_test_large_sample(s, n, t, ci = FALSE, alpha = 0.05)
}
\arguments{
\item{s}{number of successes}

\item{n}{sample size}

\item{t}{benchmark rate}

\item{ci}{Boolean, if TRUE adjusted Wald confidence interval is also reported}

\item{alpha}{for confidence interval}
}
\value{
probability of benchmark is exceeded or if confidence interval is also reported a
vector of length 3 (probability and two CI bounds)
}
\description{
Benchmark test for large sample
Small sample: success and failure counts are at least 15
Performs a one-sided test
}
