% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_test_small_sample.R
\name{benchmark_test_small_sample}
\alias{benchmark_test_small_sample}
\title{Benchmark test for small sample (binomial distribution is used)
Small sample: success and failure counts are below 15
Performs a one-sided test}
\usage{
benchmark_test_small_sample(s, n, t, mid = FALSE, ci = FALSE, alpha = 0.05)
}
\arguments{
\item{s}{number of successes}

\item{n}{sample size}

\item{t}{benchmark rate}

\item{mid}{Boolean, if TRUE mid-probability is reported (correction for discreteness in small samples)}

\item{ci}{Boolean, if TRUE adjusted Wald confidence interval is also reported}

\item{alpha}{for confidence interval}
}
\value{
exact or mid- probability of benchmark is exceeded or if confidence interval is also reported a
vector of length 3 (probabiliy and two CI bounds)
}
\description{
Benchmark test for small sample (binomial distribution is used)
Small sample: success and failure counts are below 15
Performs a one-sided test
}
