% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_log.R
\name{confidence_log}
\alias{confidence_log}
\title{Confidence interval for data with log-transformation (i.e. for skewed data, such as task times)
The t-distribution adjusts for the sample size
With larger sample verges to normal z-confidence interval}
\usage{
confidence_log(x, l)
}
\arguments{
\item{x}{data, numeric vector}

\item{l}{confidence level, decimal}
}
\description{
Confidence interval for data with log-transformation (i.e. for skewed data, such as task times)
The t-distribution adjusts for the sample size
With larger sample verges to normal z-confidence interval
}
