% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_percentile.R
\name{confidence_percentile}
\alias{confidence_percentile}
\title{Estimates percentile confidence interval (e.g. 0.5 means the median}
\usage{
confidence_percentile(x, p, l)
}
\arguments{
\item{x}{data, numeric vector}

\item{p}{percentile, decimal}

\item{l}{confifence level given as decimal, e.g. 95\% is 0.95}
}
\value{
numeric vector of interval of length 2
}
\description{
Estimates percentile confidence interval (e.g. 0.5 means the median
}
