% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_t.R
\name{confidence_t}
\alias{confidence_t}
\title{Confidence interval for continuous data
The t-distribution adjusts for the sample size
With larger sample verges to normal z-confidence interval}
\usage{
confidence_t(x, l)
}
\arguments{
\item{x}{data, numeric vector}

\item{l}{confidence level, decimal}
}
\description{
Confidence interval for continuous data
The t-distribution adjusts for the sample size
With larger sample verges to normal z-confidence interval
}
