% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_wald.R
\name{confidence_wald}
\alias{confidence_wald}
\title{Estimated the Wald (Laplace) confidence interval for binomial distribution}
\usage{
confidence_wald(s, n, l)
}
\arguments{
\item{s}{success count}

\item{n}{sample size}

\item{l}{confifence level given as decimal, e.g. 95\% is 0.95}
}
\value{
numeric vector of interval of length 2
}
\description{
Estimated the Wald (Laplace) confidence interval for binomial distribution
}
