% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_paired_sample.R
\name{generate_paired_sample}
\alias{generate_paired_sample}
\title{Generated paired sample}
\usage{
generate_paired_sample(n, s = 2)
}
\arguments{
\item{n}{numeric, sample size}

\item{s}{numeric, number of samples, defaults to 2}
}
\value{
list of vectors
}
\description{
Generated paired sample
}
