% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_mean.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Computes the geometric mean}
\usage{
geometric_mean(x)
}
\arguments{
\item{x}{data, numberic vector}
}
\value{
the exponential of the mean of log-transformed data
}
\description{
Computes the geometric mean
}
