% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps_estimator.R
\name{nps_estimator}
\alias{nps_estimator}
\title{Estimated NPS}
\usage{
nps_estimator(data, detail = FALSE)
}
\arguments{
\item{data}{(numeric vector of NPS ratings)}

\item{detail, }{Boolean, if TRUE returns NPS and detailed groups statistics}
}
\value{
NPS score as decimal
}
\description{
Estimated NPS
}
