% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_two_sample_t.R
\name{paired_two_sample_t}
\alias{paired_two_sample_t}
\title{Paired two sample t-test}
\usage{
paired_two_sample_t(x, y, alpha = 0.05)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{alpha}{confidence level}
}
\description{
Paired two sample t-test
}
